Update platform API to current schema
Adds APIs available as of 13 January 2017. Among other things, these include:

• Teams (organizations)
• Federated log-in and identity providers
• App formations sets
• Heroku SSL (SNI Endpoint)

These APIs include a deprecated endpoint in `PlatformAPI::Source` which is imported as `create___deprecated`. While not an ideal name, this is at least callable, and it’s not going to be mistaken for either its non-deprecated replacement (which takes a different number of parameters) or anything else that’s currently recommended.