Default to no rate-throttling
We can release this version 2.3.0 that has a warning/deprecation at the same time as a version 3.0.0 of this gem that sets a default rate throttle strategy.

The warning will look like this the first time that developers make an API request:

```
[Warning] Starting in PlatformAPI version 3+, requests will include rate throttling logic
to opt-out of this behavior set: `PlatformAPI.rate_throttle = RateThrottleClient::Null.new`
to silence this warning and opt-in to this logic, upgrade to PlatformAPI version 3+
```

If someone sets their own strategy, then they won't get this message.

If that sounds good, once this change is accepted. I'll make another PR to rev the version and update the behavior. Once that is approved we can release 2.3.0, then merge in the behavior change and release 3.0.0 shortly after.